Added project object
