+++ SOFATestsreal/UnitTest1.cs
@@ -1,4 +1,5 @@
 using Softwareontwerp_en_architectuur.Domain;
+using Softwareontwerp_en_architectuur.Domain.State;
 
 namespace SOFATestsreal
 {
