+++ Softwareontwerp en architectuur/Domain/Backlog_Item.cs
@@ -1,0 +1,50 @@
+﻿using Softwareontwerp_en_architectuur.Domain.Notifier;
+using Softwareontwerp_en_architectuur.Domain.State;
+
+namespace Softwareontwerp_en_architectuur.Domain
+{
+    public class Backlog_Item
+    {
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public string DefinitionOfDone { get; set; }
+        public IState State { get; set; } = new TodoState();
+        public Developer Developer { get; set; }
+        public List<Activity> Activities { get; set; } = new List<Activity>();
+        public List<INotifier> Notifiers { get; set; } = new List<INotifier>();
+        public DateOnly CompletedOn { get; set; }
+        public Backlog_Item(string name, string description, string definitionOfDone, Developer developer)
+        {
+            Name = name;
+            Description = description;
+            DefinitionOfDone = definitionOfDone;
+            Developer = developer;
+        }
+
+        public void Notify()
+        {
+
+        }
+
+        public void ChangeState(IState state)
+        {
+            this.State = state;
+        }
+
+        public int CountStorypoints()
+        {
+            int points = 0;
+            foreach (ICountable a in this.Activities)
+            {
+                points += a.CountStorypoints();
+            }
+            return points;
+        }
+
+        public int CountFinishedStoryPoints()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
+
