+++ Softwareontwerp en architectuur/Domain/Developer.cs
@@ -1,0 +1,11 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain
+{
+    public class Developer
+    {
+        public string Name { get; set; }
+        public string Email { get; set; }
+        public string Password { get; set; }
+        public string Role { get; set; }
+        public Developer() { }
+    }
+}
