+++ Softwareontwerp en architectuur/Domain/ISprint.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using Softwareontwerp_en_architectuur.Domain.State;
+using System.Collections;
 
 namespace Softwareontwerp_en_architectuur.Domain
 {
@@ -10,7 +11,7 @@ namespace Softwareontwerp_en_architectuur.Domain
     }
     public interface ISprint
     {
-
+        public bool IsFinished();
     }
     public class ReviewSprint : ISprint, ICountable
     {
@@ -41,6 +42,12 @@ namespace Softwareontwerp_en_architectuur.Domain
             }
             return points;
         }
+
+        public bool IsFinished()
+        {
+            return State.GetType() == typeof(DoneSprintState);
+            
+        }
     }
 
     public class ReleaseSprint : ISprint
@@ -51,186 +58,11 @@ namespace Softwareontwerp_en_architectuur.Domain
         public ISprintState State { get; set; }
 
         public List<string> PipelineSteps { get; set; }
-    }
 
-    public class Backlog_Item : ICountable
-    {
-        
-
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public string DefinitionOfDone { get; set; }
-        public IState State { get; set; } = new TodoState();
-        public Developer Developer { get; set; }
-        public List<Activity> Activities { get; set; } = new List<Activity>();
-        public List<INotifier> Notifiers { get; set; } = new List<INotifier>();
-        public DateOnly CompletedOn { get; set; }
-
-        public Backlog_Item(string name, string description, string definitionOfDone, Developer developer)
+        public bool IsFinished()
         {
-            Name = name;
-            Description = description;
-            DefinitionOfDone = definitionOfDone;
-            Developer = developer;
+            return State.GetType() == typeof(DoneSprintState);
         }
-
-        public void Notify()
-        {
-
-        }
-
-        public void ChangeState(IState state)
-        {
-            this.State = state;
-        }
-
-        public int CountStorypoints()
-        {
-            int points = 0;
-            foreach (ICountable a in this.Activities) {
-                points += a.CountStorypoints();
-            }
-            return points;
-        }
-
-        public int CountFinishedStoryPoints()
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class Activity : ICountable
-    {
-        public Activity(string name, string description, Developer developer, int storyPoints)
-        {
-            Name = name;
-            Description = description;
-            Developer = developer;
-            StoryPoints = storyPoints;
-        }
-
-        public string Name { get; set;}
-        public string Description { get; set; }
-        public IState State { get; set; } = new TodoState();
-
-        public Developer Developer { get; set; }
-
-        public int StoryPoints { get; set; }
-
-        
-
-        public int CountStorypoints()
-        {
-            return StoryPoints;
-        }
-
-        public int CountFinishedStoryPoints()
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class Developer
-    {
-        public string Name { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public string Role { get; set; }
-        public Developer() { }
-    }
-
-    
-
-    public interface INotifier
-    {
-        public void SendNotification();
-    }
-
-    public class EmailNotifier : INotifier
-    {
-        public void SendNotification()
-        {
-           
-        }
-    }
-
-    public class SlackNotifier : INotifier
-    {
-        public void SendNotification()
-        {
-
-        }
-    }
-
-    public interface IState
-    {
-        public void NextState(IState state);
-
-    }
-
-    public class TodoState : IState
-    {
-
-        public void NextState(IState state)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class DoingState : IState
-    {
-        public void NextState(IState state)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    public class ReadyForTestingState : IState
-    {
-        public void NextState(IState state)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    public class TestingState : IState
-    {
-        public void NextState(IState state)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    public class TestedState : IState
-    {
-        public void NextState(IState state)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class DoneState : IState
-    {
-        public void NextState(IState state)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public interface ISprintState
-    {
-
-    }
-
-    public class CreatedSprintState : ISprintState
-    {
-
-    }
-
-    public class DoingSprintState : ISprintState
-    {
-
-    }
-    public class DoneSprintState : ISprintState
-    {
-
     }
 
 }
