+++ Softwareontwerp en architectuur/Domain/Notifier/EmailNotifier.cs
@@ -1,0 +1,10 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain.Notifier
+{
+    public class EmailNotifier : INotifier
+    {
+        public void SendNotification()
+        {
+
+        }
+    }
+}
