+++ Softwareontwerp en architectuur/Domain/Notifier/INotifier.cs
@@ -1,0 +1,7 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain.Notifier
+{
+    public interface INotifier
+    {
+        public void SendNotification();
+    }
+}
