+++ Softwareontwerp en architectuur/Domain/Notifier/SlackNotifier.cs
@@ -1,0 +1,10 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain.Notifier
+{
+    public class SlackNotifier : INotifier
+    {
+        public void SendNotification()
+        {
+
+        }
+    }
+}
