+++ Softwareontwerp en architectuur/Domain/State/ISprintState.cs
@@ -1,0 +1,6 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain.State
+{
+    public interface ISprintState
+    {
+    }
+}
