+++ Softwareontwerp en architectuur/Domain/State/IState.cs
@@ -1,0 +1,8 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain.State
+{
+    public interface IState
+    {
+        public void NextState(IState state);
+
+    }
+}
