+++ Softwareontwerp en architectuur/Domain/State/SprintStateImplementations.cs
@@ -1,0 +1,16 @@
+﻿namespace Softwareontwerp_en_architectuur.Domain.State
+{
+    public class CreatedSprintState : ISprintState
+    {
+
+    }
+
+    public class DoingSprintState : ISprintState
+    {
+
+    }
+    public class DoneSprintState : ISprintState
+    {
+
+    }
+}
